mv url hash map config override fn to its own module
