+++ src/index.js
@@ -2,6 +2,7 @@
 import "styles/index.scss";
 
 import { getAllData } from "utils/data";
+import { paseUriHash } from "utils/parse-hash";
 
 /******************************************
  * GLOBAL CONSTANTS & FLAGS
@@ -68,74 +69,7 @@ let mapConfig = {
 };
 
 // read url hash input & maybe override mapConfig props
-inputValues(location.hash);
-
-// check the url hash for params then
-// override map default settings if any are present
-// assumes params are as follows:
-// #lat=<float>&lng=<float>&z=<integer>&states=<boolean>&cities=<boolean>&counties=<boolean>&rentstrike=<boolean>
-function inputValues(hash) {
-  let input = hash.slice(1).split("&");
-  let inputVals = {};
-  let i = 0;
-  for (i; i < input.length; i++) {
-    let [key, value] = input[i].split("=");
-    inputVals[key] = value;
-  }
-
-  // override the default map config values, if they exist
-  if (!isNaN(inputVals.z)) {
-    mapConfig.z = parseInt(inputVals.z);
-  }
-
-  if (!isNaN(inputVals.lat)) {
-    mapConfig.lat = parseFloat(inputVals.lat);
-  }
-
-  if (!isNaN(inputVals.lng)) {
-    mapConfig.lng = parseFloat(inputVals.lng);
-  }
-
-  if (inputVals.cities !== undefined) {
-    if (inputVals.cities === "true") {
-      mapConfig.cities = true;
-    } else if (inputVals.cities === "false") {
-      mapConfig.cities = false;
-    }
-  }
-
-  if (inputVals.counties !== undefined) {
-    if (inputVals.counties === "true") {
-      mapConfig.counties = true;
-    } else if (inputVals.counties === "false") {
-      mapConfig.counties = false;
-    }
-  }
-
-  if (inputVals.states !== undefined) {
-    if (inputVals.states === "true") {
-      mapConfig.states = true;
-    } else if (inputVals.states === "false") {
-      mapConfig.states = false;
-    }
-  }
-
-  if (inputVals.nations !== undefined) {
-    if (inputVals.nations === "true") {
-      mapConfig.nations = true;
-    } else if (inputVals.nations === "false") {
-      mapConfig.nations = false;
-    }
-  }
-
-  if (inputVals.rentstrike !== undefined) {
-    if (inputVals.rentstrike === "true") {
-      mapConfig.rentStrikes = true;
-    } else if (inputVals.rentstrike === "false") {
-      mapConfig.rentStrikes = false;
-    }
-  }
-}
+paseUriHash(mapConfig);
 
 // create a new map instance by referencing the appropriate html element by its "id" attribute
 const map = L.map("map", mapOptions).setView(
